# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TatkalDoctor Doctor Listing API",
        "description": "Core registry API for the TatkalDoctor ecosystem.\n\n## Authentication\n\n### Public Endpoints\n`GET /api/v1/health` is public — no authentication required.\n\n### Admin Login (Sanctum)\nCall `POST /api/v1/login` with email + password to receive a Bearer token.\n\n### HMAC-SHA256 (All Data Endpoints)\nAll master data and listing endpoints require four request headers:\n\n| Header | Description |\n|---|---|\n| `X-Api-Key` | Client API key issued by super admin |\n| `X-Timestamp` | Unix timestamp in seconds (UTC) |\n| `X-Nonce` | Random string ≥ 8 chars, unique per request |\n| `X-Signature` | HMAC-SHA256 hex-encoded signature |\n\n**String to sign** (fields joined by `\\n`):\n\n```\nHTTP_METHOD\nTIMESTAMP\nNONCE\nREQUEST_PATH\nSHA256(body)\n```\n\n> Timestamp must be within **±5 minutes** of server time. The `REQUEST_PATH` is the URL path only — no query string. For GET requests, `SHA256(body)` is the SHA-256 of an empty string.",
        "contact": {
            "name": "TatkalDoctor Admin",
            "email": "superadmin@tatkaldoctor.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server (WAMP)"
        }
    ],
    "paths": {
...
        "/api/v1/listings/search": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Search active doctor listings",
                "description": "Returns paginated active, verified doctor listings. Supports filtering by location, service, qualification and free-text search. Requires HMAC authentication.",
                "operationId": "searchListings",
                "parameters": [
                    {
                        "name": "country_code",
                        "in": "query",
                        "description": "Filter by ISO alpha-3 country code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "IND"
                        }
                    },
                    {
                        "name": "master_city_id",
                        "in": "query",
                        "description": "Filter by city ID (from /cities/{country_code})",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "master_location_id",
                        "in": "query",
                        "description": "Filter by location ID (from /locations/{city_id})",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "description": "Filter by service ID (from /services). Matches both parent and child services.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Free-text search on doctor name or hospital name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Ravi",
                            "maxLength": 191
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Results per page (1–100, default 20)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20,
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (default 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1,
                            "default": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listings fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Listings fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ListingSearch"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid HMAC headers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error (invalid parameter values)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HmacApiKey": [],
                        "HmacTimestamp": [],
                        "HmacNonce": [],
                        "HmacSignature": []
                    }
                ]
            }
        },
        "/api/v1/listings/slug/{qrSlug}": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Get doctor profile by QR slug",
                "description": "Returns a full active, verified doctor profile by its QR slug. Used by public profile pages accessible via QR code scan. This route must be called **before** `/listings/{uuid}` to avoid parameter collision. Requires HMAC authentication.",
                "operationId": "showListingBySlug",
                "parameters": [
                    {
                        "name": "qrSlug",
                        "in": "path",
                        "description": "Unique QR slug generated when a listing is approved (e.g. dr-ravi-kumar-550e8400)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "dr-ravi-kumar-550e8400"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listing fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Listing fetched successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ListingDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid HMAC headers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Listing not found (no active approved listing with this slug)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Listing not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HmacApiKey": [],
                        "HmacTimestamp": [],
                        "HmacNonce": [],
                        "HmacSignature": []
                    }
                ]
            }
        },
        "/api/v1/listings/{uuid}": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Get doctor profile by UUID",
                "description": "Returns a full active, verified doctor profile by its UUID. The response includes all profile fields, contact details, qualifications, services and QR data. Requires HMAC authentication.",
                "operationId": "showListing",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the doctor listing",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listing fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Listing fetched successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ListingDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid HMAC headers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Listing not found (no active approved listing with this UUID)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Listing not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HmacApiKey": [],
                        "HmacTimestamp": [],
                        "HmacNonce": [],
                        "HmacSignature": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "description": "Standard error response body",
                "properties": {
                    "success": {
                        "description": "Reusable OpenAPI schema definitions.\nReferenced via $ref: '#/components/schemas/{name}'.",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message describing what went wrong"
                    }
                },
                "type": "object"
            },
...
            "ListingSearch": {
                "description": "Doctor listing — compact search result item",
                "properties": {
                    "uuid": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Dr. Ravi Kumar"
                    },
                    "hospital_name": {
                        "type": "string",
                        "example": "City General Hospital",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "example": "Mumbai",
                        "nullable": true
                    },
                    "location": {
                        "type": "string",
                        "example": "Andheri",
                        "nullable": true
                    },
                    "average_rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.8
                    }
                },
                "type": "object"
            },
            "ListingDetail": {
                "description": "Full doctor listing profile",
                "properties": {
                    "uuid": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Dr. Ravi Kumar"
                    },
                    "hospital_name": {
                        "type": "string",
                        "example": "City General Hospital",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "123 MG Road, Mumbai",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Experienced cardiologist with 15+ years practice.",
                        "nullable": true
                    },
                    "country": {
                        "type": "string",
                        "example": "India",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "example": "Mumbai",
                        "nullable": true
                    },
                    "location": {
                        "type": "string",
                        "example": "Andheri",
                        "nullable": true
                    },
                    "personal_contact_no": {
                        "type": "string",
                        "example": "+91 9000000000",
                        "nullable": true
                    },
                    "appointment_no": {
                        "type": "string",
                        "example": "+91 9000000001",
                        "nullable": true
                    },
                    "qualifications": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "MBBS",
                            "MD — Cardiology"
                        ]
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Cardiology",
                            "Interventional Cardiology"
                        ]
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 19.076,
                        "nullable": true
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 72.8777,
                        "nullable": true
                    },
                    "average_rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.8
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "qr_slug": {
                        "type": "string",
                        "example": "dr-ravi-kumar-550e8400",
                        "nullable": true
                    },
                    "public_profile_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://tatkaldoctors.com/d/dr-ravi-kumar-550e8400",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "description": "Pagination metadata",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "HmacApiKey": {
                "type": "apiKey",
                "description": "Your client API key.",
                "name": "X-Api-Key",
                "in": "header"
            },
            "HmacTimestamp": {
                "type": "apiKey",
                "description": "Current Unix timestamp (UTC). Must be within ±5 minutes of server time.",
                "name": "X-Timestamp",
                "in": "header"
            },
            "HmacNonce": {
                "type": "apiKey",
                "description": "A random string (min 8 chars) unique per request.",
                "name": "X-Nonce",
                "in": "header"
            },
            "HmacSignature": {
                "type": "apiKey",
                "description": "HMAC-SHA256 hex signature. String to sign: \"METHOD\\nTIMESTAMP\\nNONCE\\nPATH\\nSHA256(body)\".",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
...
}